refactor(cli): Phase 2a of cli/main.py split — connectors + init + clean + serve + _helpers
Continuation of the cli/main.py extraction work started in #565
(Phase 1 — doctor / list_syncs / config / cloud / docs / mcp). Same
pattern: each command lives in its own module under
``drt/cli/commands/``, decorators register on the shared ``app`` from
``drt.cli._app``.

What moves
----------

New ``drt/cli/_helpers.py`` — shared CLI internals:

- ``resolve_profile_name(cli_flag, project_profile)``
- ``get_source(profile)`` — wraps ``drt.connectors.get_source``
- ``get_destination(sync)`` — wraps ``drt.connectors.get_destination``
- ``get_watermark_storage(sync, project_dir)`` — branches on
  ``wm.storage`` to instantiate ``LocalWatermarkStorage`` /
  ``GCSWatermarkStorage`` / ``BigQueryWatermarkStorage``

Five new modules under ``drt/cli/commands/``:

- ``connectors.py`` — ``drt sources`` + ``drt destinations`` (paired
  because they share the Rich-table / Rich-panel / JSON output
  helpers from #543)
- ``clean.py`` — ``drt clean --orphans``
- ``serve.py`` — ``drt serve`` (HTTP webhook trigger)
- ``init.py`` — ``drt init`` with all three modes (interactive
  wizard / ``--template <name>`` / ``--from-dbt <manifest>``) and
  their private helpers (``_list_templates``,
  ``_init_from_template``, ``_init_from_dbt``)

``commands/__init__.py`` registers the five new modules alongside
the six from Phase 1.

main.py changes
---------------

- Removes the five moved commands and their inline helpers (the
  three connector printers, the three init helpers) — replaced with
  one-line pointer comments
- ``_get_source`` / ``_get_destination`` / ``_get_watermark_storage``
  are now **thin back-compat shims** delegating to
  ``drt.cli._helpers.*``. Preserves the ``from drt.cli.main import
  _get_destination`` import path that tests rely on (e.g. the
  connector registry tests still patch via the main module).

main.py: **1468 → 1083 LOC** (-385, -26%). Cumulative since Phase 1:
1706 → 1083 (-36%).

Test impact
-----------

``tests/unit/test_clean_command.py``: 4 patch sites updated from
``drt.cli.main._get_destination`` to
``drt.cli.commands.clean.get_destination`` — the clean command now
imports the factory from ``_helpers`` directly, so the old patch
target no longer intercepts the call. (The back-compat shim on
main.py is still present for tests that import the symbol directly
rather than patching it.)

All other tests pass unchanged — the shims cover the rest.

Out of scope (Phase 2b — separate follow-up)
--------------------------------------------

Heavier commands stay in main.py for a dedicated PR with focused
review attention:

- ``run`` + ``_RunContext`` + ``_run_one`` + signal handling +
  threading (~450 LOC, the riskiest of all the remaining commands)
- ``validate`` (~190 LOC, secrets scan + connection check)
- ``status`` (~120 LOC, StateManager interplay)
- ``test`` + ``_SyncTestResult`` (~130 LOC, sync test framework)
- ``_JsonFormatter`` + ``_configure_json_logging`` (~25 LOC, JSON
  logging infrastructure shared by ``run``)

Each is mechanical now — the pattern is fully proven across 11
commands moved. Phase 2b can land as one PR or be split further by
command. Filing as a separate issue.

Local verification
------------------

  pytest tests/                  → 1218 passed, 1 skipped, 7 deselected
  ruff check drt tests           → All checks passed
  mypy drt                       → Success: no issues in 109 files

Refs: #546 (the umbrella issue, Phase 1 closed it but Phase 2 work
continues — the issue body acknowledges multi-PR scope), #565 (the
pattern this PR extends), #538 (Tech Foundation Hardening epic).
